pyproject.toml
[tool.poetry]
name = "pjecz-hermes"
version = "0.1.0"
description = "Hermes es el encargado de mandar los mensajes con documentos y de llevar una bitácora de destinatarios."
authors = ["Guillermo Valdes <guillermo.valdes@pjecz.gob.mx>"]
license = "AGPL3"
readme = "README.md"
packages = [{include = "pjecz_hermes"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
redis = "^4.6.0"
rich = "^13.4.2"
sendgrid = "^6.10.0"
typer = "^0.9.0"
unidecode = "^1.3.6"


[tool.poetry.group.dev.dependencies]
faker = "^18.13.0"
black = "^23.3.0"
pylint = "^2.17.4"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 256
